pyproject.toml
[project]
name = "discord-summary"
version = "0.1.0"
description = "Discord chat history exporter with daily markdown snapshots"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
discord-summary = "discord_summary.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_summary"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
